Changed setid to correctly be passed a long instead of an int

+++ On Your Bike - Chapter 11/src/com/androiddevbook/onyourbike/chapter11/model/Trip.java
@@ -67,7 +67,7 @@ public class Trip {
         database.execSQL(dropTable);
     }
 
-    public void setid(int id) {
+    public void setid(long id) {
         _id = id;
     }
